add support for Python 3.12 and drop EoL Python 3.7
``astroid`` 3.x is required to support Python 3.12, and it does not
support Python 3.7. Instead of installing an older version of
``astroid`` Python 3.7 was dropped because it already reached its end of
life.

In order to remove deprecation warnings other development dependencies
have been updated to their latest versions and pinned packages have been
set to their minor versions instead of pinning to the patch versions.